v0.7.2: Creating a repeater class to track time in the program and perform repetitive actions, the ability to pass a non-integer number of seconds in the speed calculation

+++ maid_module/devices.h
@@ -19,7 +19,7 @@ private:
     uint impulse_counter;
 public:
     Speedometer(uint impulse_cnt);
-    float CalculateSpeed(byte time_spent_sec, Wheel &wheel);
+    float CalculateSpeed(float time_spent_sec, Wheel &wheel);
     float CalculateAverageSpeed(Node* head);
     void CountImpulse();
     void ResetCounter();
